requirements.txt
# Core Dependencies
pandas==2.0.3
numpy==1.24.3
python-dateutil==2.8.2

# Data Collection & Trading APIs
ccxt==4.1.0  # Cryptocurrency exchanges
requests==2.31.0

# Database
sqlalchemy==2.0.19

# Technical Analysis
ta==0.10.2  # Technical analysis library

# Machine Learning (for ML strategy)
scikit-learn==1.3.0
xgboost==2.0.0

# Configuration Management
omegaconf==2.3.0
python-dotenv==1.0.0
pyyaml==6.0.1

# Visualization & Monitoring
plotly==5.15.0
dash==2.14.0
matplotlib==3.7.2

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Logging
colorlog==6.7.0

# Async Support
asyncio
aiohttp==3.9.1

# Utilities
tqdm==4.66.1  # Progress bars
